backend/requirements.txt
fastapi
uvicorn
sqlalchemy
sqlmodel
supabase
pydantic
python-dotenv
reportlab
python-multipart
passlib[bcrypt]
